Refactor[game]: reorganize imports in page and index files
