+++ src/lib/features/game/index.ts
@@ -1,2 +1,8 @@
+import Field from "./components/Field/Field";
+import Info from "./components/Info/Info";
+import { useGameChecker } from "./hooks/useGameChecker";
+
 export * from "./game";
-export * from "../../../app/game/page";
+export { Field };
+export { Info };
+export { useGameChecker };
